Renamed space_modeling to map
