+++ tests/test_map.py
@@ -1,5 +1,5 @@
 import pytest
-from src.space_modeling import Map
+from src.map import Map
 
 @pytest.fixture
 def random_map():
